refactor: 유저 삭제 확인 도메인 로직으로 변경
