+++ aics-auth/src/main/java/kgu/developers/auth/api/application/AuthService.java
@@ -1,6 +1,5 @@
 package kgu.developers.auth.api.application;
 
-import kgu.developers.auth.api.presentation.exception.AlreadyDeletedUserException;
 import kgu.developers.auth.api.presentation.exception.TokenNotFoundException;
 import kgu.developers.auth.api.presentation.request.LoginRequest;
 import kgu.developers.auth.api.presentation.request.RefreshTokenRequest;
@@ -31,10 +30,7 @@ public class AuthService {
 		String password = request.password();
 
 		User user = userQueryService.getUserById(userId);
-
-		if (user.getDeletedAt() != null)
-			throw new AlreadyDeletedUserException();
-
+		user.isDeleted();
 		user.isPasswordMatching(password, passwordEncoder);
 
 		String role = user.getRole().name();
