+++ aics-domain/src/main/java/kgu/developers/domain/user/exception/AlreadyDeletedUserException.java
@@ -1,8 +1,8 @@
-package kgu.developers.auth.api.presentation.exception;
+package kgu.developers.domain.user.exception;
 
 import kgu.developers.common.exception.CustomException;
 
-import static kgu.developers.auth.api.presentation.exception.AuthExceptionCode.ALREADY_DELETED_USER;
+import static kgu.developers.domain.user.exception.UserDomainExceptionCode.ALREADY_DELETED_USER;
 
 public class AlreadyDeletedUserException extends CustomException {
 	public AlreadyDeletedUserException() {
