+++ aics-domain/src/main/java/kgu/developers/domain/user/exception/UserDomainExceptionCode.java
@@ -22,6 +22,7 @@ public enum UserDomainExceptionCode implements ExceptionCode {
 	USER_NOT_FOUND(NOT_FOUND, "해당 회원을 찾을 수 없습니다."),
 	USER_ID_DUPLICATED(CONFLICT, "이미 동일한 학번으로 가입이 되어있습니다."),
 	NOT_DELETABLE_USER(BAD_REQUEST, "일반 사용자만 삭제 가능합니다."),
+	ALREADY_DELETED_USER(BAD_REQUEST, "이미 탈퇴한 회원입니다.")
 	;
 
 	private final HttpStatus status;
